# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Picroom API
  version: 1.0.0
  description: |
    Picroom — self-hosted image hosting service.

    Two API surfaces:
    1. REST/JSON under `/api/v1/` — first-class management API.
    2. AWS S3-compatible under `/s3/` — drop-in replacement for object storage
       clients (PicGo, rclone, AWS CLI).

    All endpoints require authentication except where noted.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Picroom Maintainers
    url: https://github.com/picroom/picroom
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://{host}
    description: Self-hosted production
    variables:
      host:
        default: picroom.example.com
...
paths:
...
  /api/v1/images:
    get:
      tags: [images]
      summary: List images
      operationId: listImages
      parameters:
        - in: query
          name: team_id
          schema:
            type: string
            format: uuid
        - in: query
          name: owner_id
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImagePage'
    post:
      tags: [images]
      summary: Upload an image (multipart)
      operationId: uploadImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                team_id:
                  type: string
                  format: uuid
                storage_policy:
                  type: string
                  default: default
              required: [file]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: Quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '415':
          description: Unsupported media type
  /api/v1/images/{imageId}:
    parameters:
      - $ref: '#/components/parameters/ImageId'
    get:
      tags: [images]
      summary: Get image metadata
      operationId: getImage
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [images]
      summary: Delete an image
      operationId: deleteImage
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/images/{imageId}/file:
    parameters:
      - $ref: '#/components/parameters/ImageId'
    get:
      tags: [images]
      summary: Get original image bytes (redirect to public URL)
      operationId: getImageFile
      responses:
        '302':
          description: Redirect to the public /i/{key} URL
          headers:
            Location:
              schema:
                type: string
                format: uri
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/images/{imageId}/link:
    parameters:
      - $ref: '#/components/parameters/ImageId'
    get:
      tags: [images]
      summary: Get the public link ("公链") for an image
      operationId: getImageLink
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [public_url]
                properties:
                  public_url:
                    type: string
                    description: Absolute URL when server.public_url_base is set, else path-relative.
                  expires_at:
                    type: string
                    format: date-time
                    nullable: true
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /i/{key}:
    parameters:
      - in: path
        name: key
        required: true
        schema:
          type: string
        description: Storage key (e.g. `img/{uuid}.bin`), may contain slashes.
    get:
      tags: [images]
      summary: Public image bytes (no auth, "公链")
      description: |
        Serves raw object bytes without authentication. Knowing the key grants
        read access (keys are unguessable UUIDs). Content-Type is sniffed from
        magic bytes. Cached immutably (1 year).
      operationId: getPublicObject
      security: []
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: picroom_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sigv4:
      type: apiKey
      in: header
      name: Authorization
      description: AWS Signature Version 4
  parameters:
    ImageId:
      in: path
      name: imageId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples: [quota_exceeded, validation_failed]
        message:
          type: string
          description: Human-readable message
        request_id:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
...
    Image:
      type: object
      required: [id, owner_id, content_type, bytes, width, height, created_at]
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
          nullable: true
        storage_policy:
          type: string
        key:
          type: string
        content_type:
          type: string
        bytes:
          type: integer
          format: int64
        width:
          type: integer
        height:
          type: integer
        sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
        variants:
          type: array
          items:
            type: string
            enum: [avif, webp, thumbnail_200, thumbnail_400, thumbnail_800]
        created_at:
          type: string
          format: date-time
    ImagePage:
      type: object
      required: [items, has_more]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
...
security:
  - cookieAuth: []
  - bearerAuth: []